% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mewAvg-package}
\alias{mewAvg-package}
\title{A fixed memory moving expanding window average}
\description{
This package provides the tools to calculate an
average in a moving expanding window (MEW) using a fixed amount of
memory.
}
\details{
See the examples for the functions \code{mewMean} and
\code{mewAvg} for the details of use.
}
\references{
Still to be added
}

