% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mewInit}
\alias{mewInit}
\title{Create an S4 object of class \code{mewTyp}}
\usage{
mewInit(n_bin, n_xx, ff)
}
\arguments{
\item{n_bin}{(scalar integer) The fixed number of bins to use to
define the moving expanding window}

\item{n_xx}{(scalar integer) The length of each vector in the
sequence to be averaged}

\item{ff}{(scalar double) The fraction of the samples to included
in each window}
}
\value{
An initialized instance of the class \code{mewTyp}
}
\description{
Call this function to create an S4 object of class
\code{mewTyp}.
}
\details{
If it is necessary to directly call mewAccum and mewMean
an S4 object of class \code{mewTyp} should be created first using
this function.  The user should never create an S4 object of class
\code{mewTyp} using the \code{new} function provided by the
\code{methods} package.
}
\examples{
av <- mewInit(n_bin = 4, n_xx = 2, ff = 0.5)
}

