% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,mewTyp-method}
\alias{show,mewTyp-method}
\title{Print the MEW average to the screen}
\usage{
\S4method{show}{mewTyp}(object)
}
\arguments{
\item{object}{(class mewTyp) The current state of the MEW average}
}
\value{
Upon successful exit, zero is returned invisibly.
}
\description{
Print to the screen the first six elements of the
current value (if it is up-to-date) of the moving expanding window
(MEW) average. An error is raised if the MEW average is not
up-to-date.
}
\examples{
## see the examples for the function mewMean
}

