% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mewAvg.R
\name{mewAvg-package}
\alias{mewAvg-package}
\title{A fixed memory moving expanding window average}
\description{
This package provides the tools to calculate an
average in a moving expanding window (MEW) using a fixed amount of
memory.
}
\details{
See the examples for the functions \code{mewMean} and
\code{mewAvg} for the details of use.
}
\references{
Levine, Z. H., & Pintar, A. L. (2015). A fixed-memory moving,
expanding window for obtaining scatter corrections in x-ray CT and other
stochastic averages. Computer Physics Communications, 196, 455-459.
}
